requirements.txt
Flask==2.3.3
flask-cors==4.0.0
flask-jwt-extended==4.5.2
Flask-SQLAlchemy==3.0.5
flask-swagger-ui==4.11.1
marshmallow==3.20.1
pyjwt==2.8.0
python-dotenv==1.0.0
google-cloud-storage==2.10.0
Werkzeug==2.3.7
gunicorn==21.2.0
pymysql==1.1.0
cryptography==41.0.3
